pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "real-time-agent"
version = "0.1.0"
description = "Reference design for real-time agents built on Azure OpenAI or AI Foundry"
readme = "README.md"
license = {file = "LICENSE"}
authors = [
    {name = "Perfect Thyme Tech", email = "info@perfectthyme.tech"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
requires-python = ">=3.8"
dependencies = [
    "openai>=1.0.0",
    "azure-identity>=1.15.0",
    "azure-ai-inference>=1.0.0",
    "websockets>=12.0",
    "asyncio-mqtt>=0.16.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "aiohttp>=3.9.0",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/PerfectThymeTech/real-time-agent"
Repository = "https://github.com/PerfectThymeTech/real-time-agent"
Issues = "https://github.com/PerfectThymeTech/real-time-agent/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ['py38']

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true